Building a separate page to test the import functionality
Based on the Bechance tutorial using handle bars and query as tools to
connect to and import relevant data from bechance. I want to build just
the portfolio elements on this page first before trying to place this
into the bechance .moving box on the index page so I don’t get confused
with what is already there.

Depending I may simply copy and paste this into the index page, or
possible use an include script…or load the page into the class another
way. Thinking include right now but we’ll see.